Add 'standalone' and 'integrated' feature flags
