+++ bee-api/bee-rest-api/Cargo.toml
@@ -14,7 +14,7 @@ homepage = "https://www.iota.org"
 bee-common = { git = "https://github.com/iotaledger/bee.git", branch = "dev", optional = true }
 bee-ledger = { path = "../../bee-ledger" }
 bee-message = { git = "https://github.com/iotaledger/bee.git", branch = "dev" }
-bee-network = { path = "../../bee-network", optional = true }
+bee-network = { path = "../../bee-network", default-features = false, optional = true }
 bee-pow = { git = "https://github.com/iotaledger/bee.git", branch = "dev" }
 bee-protocol = { path = "../../bee-protocol" }
 bee-runtime = { git = "https://github.com/iotaledger/bee.git", branch = "dev", optional = true  }
