--- bee-network/src/host/mod.rs
@@ -1,13 +1,0 @@
-// Copyright 2020 IOTA Stiftung
-// SPDX-License-Identifier: Apache-2.0
-
-// NOTE: We want things to be organized in as small as possible files, but still keep things simple
-// when importing types by making everything available through the `host` module.
-mod host;
-pub use host::*;
-
-mod error;
-pub use error::*;
-
-mod info;
-pub use info::*;
