+++ bee-network/src/peer/mod.rs
@@ -1,14 +1,9 @@
 // Copyright 2020 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-#[cfg(feature = "node")]
-mod ban;
-#[cfg(feature = "node")]
-mod error;
-mod info;
-#[cfg(feature = "node")]
-mod peerlist;
+#![cfg(feature = "standalone")]
 
-#[cfg(feature = "node")]
-pub use self::{ban::*, error::*, info::peerstate::PeerState, peerlist::*};
-pub use info::{PeerInfo, PeerRelation};
+pub mod ban;
+pub mod error;
+pub mod meta;
+pub mod store;
