+++ bee-network/src/service/command.rs
@@ -1,10 +1,9 @@
 // Copyright 2020 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::peer::PeerRelation;
+use crate::types::PeerRelation;
 
 use libp2p::{Multiaddr, PeerId};
-
 use tokio::sync::mpsc;
 
 pub type CommandReceiver = mpsc::UnboundedReceiver<Command>;
