+++ bee-network/src/swarm/mod.rs
@@ -1,10 +1,8 @@
 // Copyright 2020 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+#![cfg(feature = "standalone")]
+
+pub mod behavior;
+pub mod builder;
 pub mod protocols;
-
-mod builder;
-pub use builder::*;
-
-mod behavior;
-pub use behavior::*;
